create github section
