+++ app/core/components/github/GitHubCollaborators/index.tsx
@@ -1,0 +1,7 @@
+export default function GitHubCollaborators({
+  repoName,
+}: {
+  repoName: string;
+}) {
+  return <></>;
+}
